Be more careful when moving head/tail pointers
If we are at the edge of the board, we should not assume we intend to
move our head/tail to the other edge of the board unless we are really
moving in that direction.

Hopefully fixes the invisible positions of death that get left behind
sometimes when we go off the edge of the map or use teleporters.

https://bugzilla.gnome.org/show_bug.cgi?id=654072