Don't skip last space on board when moving left/up
This would be a harmless, hard to notice visual glitch, if not that our
test for whether the head can move to this position (different from this
code, which controls the actual movement...) thinks this doesn't happen.
So we can end of slicing off part of an enemy worm without getting
killed.

https://bugzilla.gnome.org/show_bug.cgi?id=654072