Fix corner-case handling of tail teleportation
What we do here is really insane. After our tail moves into the
teleporter, we erase it from the text board and remove its clutter
actor, but then intentionally leave the tail pointer stationary,
trusting to the gnibbles_worm_get_tail_direction function to position it
properly on the next turn, then warping and moving the tail pointer an
extra time on that turn. This kind of works because
gnibbles_worm_get_tail_direction() operates on the clutter actors, which
are stored in a linked list and do know their position. But it's also
super fragile if we do anything interesting with the teleporters (e.g.
teleport immediately on the next move, because we came at a teleporter
from a particular direction). So instead of this
don't-move-then-move-twice nonsense, assume that if we ever do not know
which direction to move in it's because we're at a teleporter, and try
to use it.

Also, add a couple sensible assertions to make sure this assumption
always holds.

https://bugzilla.gnome.org/show_bug.cgi?id=654072