Use reverse-DNS notation for appdata and desktop file
I need to hardcode the prgname in libgames-support, to handle score
import, and the prgname needs to match the desktop file name. I don't
want to hardcode a name that we know will change in the future, so
change it now.

Note to anyone reading this commit: when renaming a desktop file, you
have to update the rename list in gnome-shell/js/ui/appFavorites.js. And
you have to rename the appdata file at the same time, even though you
really shouldn't need to, given that the appdata file has a field for
which to identify the desktop file....

And if I wasn't sufficiently clear up above, make sure the prgname
matches the desktop file! GLib sets it to match the name of the
application binary by default, which is not going to work if the desktop
file name does not match the binary anymore.