Get rid of special handling for GTK+ and Clutter command line options
Nowadays the best practice is to ignore GTK+ command line options and
just use environment variables instead. (This is what GtkApplication
does; it doesn't actually pass any of the options to gtk_init. It's a
bit buggy, though, as the options still appear in help for some reason,
but that's not our fault.)

Treat the Clutter options the same way. If these options provide any
functionality not accessible via environment variables, Clutter should
be enhanced.

Accordingly, move this functionality out of main, and into startup()
where it really belongs.

https://bugzilla.gnome.org/show_bug.cgi?id=762054