Allow selecting the number of computer players from the 'Number of Players' screen
We decided that it's best that we we have a minimum number (4) of
worms in a game. Therefore, the minimum number of AI should be
the maximum number of possible worms minus the number of human players.
Knowing this, we only present buttons to select the number of AI players
based on the number of human worms selected by the user.

https://bugzilla.gnome.org/show_bug.cgi?id=762042