Handle key pressed event in gnome-nibbles.vala
The reason I am moving this from nibbles-view.vala to gnome-nibbles.vala
is to avoid a weird behavior of having your first key press ignored
everytime by the start of a new level, thus making your worm
unresponsive to your command.

As unlikey as it may seem that there is any difference of behavior
by connecting to the signal in gnome-nibbles.vala rather than
nibbles-view.vala, this actually provides a fix for the bug that states
a worm is uncontrollable briefly after spawn.

https://bugzilla.gnome.org/show_bug.cgi?id=758467