Apply bonus type HALF based on current length
When reducing the worm's length to half, the total length was actually
the sum of the worm's length and the length that was going to be earned/
lost if the worm previously consumed another bonus.

Reduce the worm's length based on the length shown when the HALF
bonus is consumed.

https://bugzilla.gnome.org/show_bug.cgi?id=762835