Make 'Next Level' button the default widget
Currently, when a level is completed, the focus is set on the Next
Level button so by pressing Enter the user can immediately start the
next level. This can sometimes lead to problems if the focus is changed.

Instead of relying on the widget's focus, make the 'Next Level' button
the default widget so by pressing enter with no focus, the next level
is always started.

https://bugzilla.gnome.org/show_bug.cgi?id=773049