app: Make app ID consistent with data files
The app ID used by .desktop and .appdata files is currently
inconsistent with the app ID claimed on D-Bus. This isn't a
problem if the application is installed through normal
distribution channels, however flatpak requires a single
consistent ID for all exported files. Some of this can
be tweaked when building a flatpak, but it's much easier
to just use the same ID from the get-go.

https://bugzilla.gnome.org/show_bug.cgi?id=780121