Fix worm position
A worm’s position was stored in two data structures, a linked list and
an array (the array also stores other data). The two grew more
inconsistent with time. By only using the linked list to store the
worms’ position the code that attempted to keep these two data store
consistent could be removed and the code that queried the worms’
position on the array changed to use the linked lists. Resulting in
much simpler working code.