Correct worm AI
Redesign of ai_move so its not possible to trigger the assert within.
Remove ai_wander (a brute force recursive function for finding a bonus)
and replace it with a more intelligent design. The code now looks for
a line of sight to bonuses only after we determine a bonus is in the
direction we are looking. Choose the closest bonus to head for
preferring LIFE bonuses.