Bug fixes and hamburger menu position change
The window size was being stored at 0 width and 0 height when the
program was closed.
Fixes a bug with the cursor keys navigating around the menu bar and
not directing your worm.
Changes the position of the hamburger menu to be more complient with
the Gnome standard.
Revert to the modal New Game dialog.