Remove deprecate warning
The use of show_uri is deprecated since GTK 4.10. The suggested alternative
is to use UriLauncher, however for help UriLauncher has a bug that results
in a "Document Not Found" message when used in flatpack.
There is no intention to fix this bug so we will continue to use show_uri.
This stops the deprecate warning and hopefully stops anyone mistakenly
coding a switch to UriLauncher.