Null terminate array of strings
When a const string [] is passed to a C function it is passed as a pointer to a list of pointers without a null terminator at the end of the list. Therefore we must manually put a null at the end of the list if the C function is expecting it (which is usually the case for a list of pointers of unknown length).