Silence C compiler warnings
Vala generated code is not intended to be built with warnings enabled.
There is currently a massive spam of C compiler warnings that are
useless because only Vala compiler developers can do anything about
them. Suppress them so we are more likely to notice real warnings from
valac.

I had already added -w to the Autotools build long ago, but it must have
gotten lost in the port to Meson.